Move Variable to variables module
